xen/arm: SMMU: Introduce a helper to add DT device to SMMU
In current code, we only have the iommu_add_device to add PCI device
to IOMMU. But for ARM SMMU, we don't have a separate helper to add
platform device with device tree to SMMU. This work was included in
the iommu_assign_dt_device. But sometimes, we just want to add device
to SMMU to do some preparation for further use. In this case, we can't
call iommu_assign_dt_device.

In previous patch, we have implement the add_device callback for SMMU,
so we can separate this work from assign_device now.

Signed-off-by: Wei Chen <Wei.Chen@arm.com>