ARM: vGIC: rework vgic_get_target_vcpu to take a pending_irq
For now vgic_get_target_vcpu takes a VCPU and an IRQ number, because
this is what we need for finding the proper rank and the VCPU in there.
In the future the VCPU will be looked up in the struct pending_irq.
To avoid locking issues, let's pass the pointer to the pending_irq
instead. We can read the IRQ number from there, and all but one caller
know that pointer already anyway.
This simplifies future code changes.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>