x86/vmx: add struct vmx_msr_policy
This structure provides a convenient way of accessing contents of
VMX MSRs: every bit value is accessible by its name. Bit names match
existing Xen's definitions as close as possible. The structure also
contains the bitmap of available MSRs since not all of them may be
available on a particular H/W.

A set of helper functions is introduced to provide a simple way of
interacting with the new structure.

Signed-off-by: Sergey Dyasli <sergey.dyasli@citrix.com>