x86/sysctl: Implement XEN_SYSCTL_get_cpuid_policy
Provide a SYSCTL for the toolstack to obtain complete system cpuid policy
information.  The CPUID information is serialised as an array of 6x 32bit
integers, and a mechanism is provided to query the maximum number of entries
Xen might write.

For the XSM side of things, this subop is closely related to
{phys,cputopo,numa}info, so shares the physinfo access vector.

Signed-off-by: Andrew Cooper <andrew.cooper3@citrix.com>