xen: move XENMAPSPACE_grant_table code into grant_table.c
The x86 and arm versions of XENMAPSPACE_grant_table handling are nearly
identical. Move the code into a function in grant_table.c and add an
architecture dependant hook to handle the differences.

This at once fixes a bug in the arm code which didn't unlock the grant
table in error case.

Signed-off-by: Juergen Gross <jgross@suse.com>