xen: support different gnttab_max_frames for grant v1 and v2
The number of grants a domain can setup is limited by the maximum
number of grant frames it is allowed to use. Today the limit is the
same regardless whether the domain uses grant v1 or v2. Using v2
will therefor be a disadvantage for the domain as only half the
number of grants compared to v1 can be used, because a grant v2 entry
is twice as large as the v1 entry. This is the reason for the lack of
grant v2 support in the Linux kernel (in fact grant v2 support has
been removed from Linux for this reason).

OTOH using only grant v1 will limit a pv domain to the low 16TB of
memory of the host, as grant v1 entries only use a 32 bit mfn. So
if we want to support more than 16TB of memory and be able to use
that memory in pv domains, we have to remove the disadvantage of
using grant v2 by being able to setup the same number of grants as
with v1.

In order to achieve this add support for limiting the number of grant
frames for v1 and v2 independently from each other. Per default let
the v2 number be twice the value of the v1 number. Modify the boot
parameter gnttab_max_frames to accept either a single value which
will set the v1 limit to that value and the v2 limit to 2*value, or
two values separated by a comma to set both limits to dedicated
values.

Add some sanity checks to make sure the maximum number of frames isn't
lower than the initial number, as this leads to rather strange crashes.

Signed-off-by: Juergen Gross <jgross@suse.com>