x86/cpuid: emulate extended topology enumeration leaf
This patch sets policy for guest's CPUID.0xb.  eax return value can be infered
from the number of core/thread in one socket/core. ebx cannot be zero
otherwise it means CPUID.0xb isn't supported. edx is the x2apic id of each
vcpu, which would be adjusted in Xen according the context. ecx[7:0] should be
the same value in ecx input .  ecx[15:8] should be the level type.

This patch also sets policy for guest's CPUID.4:EAX[25:14]. Rather than
passing the host value to guest, we set this field to the number of thread
in each core to be consistent with CPU topology.

Signed-off-by: Chao Gao <chao.gao@intel.com>